+++ src/components/InstanceCard/InstanceCard.tsx
@@ -2,15 +2,23 @@ import React from "react";
 
 import { Instance, WorkEdition } from "~/src/types/DataModel";
 import {
+  Box,
   Card,
   CardActions,
   CardContent,
   CardHeading,
+  Flex,
 } from "@nypl/design-system-react-components";
 import EditionCardUtils from "~/src/util/EditionCardUtils";
 import Link from "../Link/Link";
-import { useCookies } from "react-cookie";
-import { NYPL_SESSION_ID } from "~/src/constants/auth";
+import Ctas from "../EditionCard/Ctas";
+import PublisherAndLocation from "../EditionCard/PublisherAndLocation";
+import WorldCat from "./WorldCat";
+import CardRequiredBadge from "../EditionCard/CardRequiredBadge";
+import FeaturedEditionBadge from "../EditionCard/FeaturedEditionBadge";
+import PhysicalEditionBadge from "../EditionCard/PhysicalEditionBadge";
+import ScanAndDeliverBlurb from "../EditionCard/ScanAndDeliverBlurb";
+import UpBlurb from "../EditionCard/UpBlurb";
 
 // Creates an Instance card out of the Edition Year and Instance object
 // Note: Edition Year only needs to be passed because `instance.publication_date`
@@ -20,50 +28,75 @@ import { NYPL_SESSION_ID } from "~/src/constants/auth";
 export const InstanceCard: React.FC<{
   edition: WorkEdition;
   instance: Instance;
+  isFeaturedEdition?: boolean;
 }> = (props) => {
-  // cookies defaults to be undefined if not fonud
-  const [cookies] = useCookies([NYPL_SESSION_ID]);
-
   const edition = props.edition;
   const instance: Instance = props.instance;
+  const isFeaturedEdition = props.isFeaturedEdition;
   const previewItem = EditionCardUtils.getPreviewItem(instance.items);
+  const isPhysicalEdition = EditionCardUtils.isPhysicalEdition(previewItem);
+  const isUniversityPress = EditionCardUtils.isUniversityPress(previewItem);
+  const isLoginRequired = isPhysicalEdition || isUniversityPress;
 
   return (
-    <Card
-      imageProps={{
-        src: EditionCardUtils.getCover(edition.links),
-        size: "xsmall",
-        aspectRatio: "original",
-        alt: "Cover",
-      }}
-      layout="row"
-      isBordered
-      isAlignedRightActions
-      id={`card-${instance.instance_id}`}
-      p="s"
+    <Box
+      border="1px"
+      borderColor="ui.border.default"
+      padding="s"
+      paddingBottom="l"
+      paddingRight="l"
     >
-      <CardHeading level="h3" size="heading6" sx={{ fontSize: "18px" }}>
-        {edition.publication_date
-          ? edition.publication_date
-          : "Edition Year Unknown"}
-      </CardHeading>
-      <CardContent>
-        <div>
-          {EditionCardUtils.getPublisherAndLocation(
-            instance.publication_place,
-            instance.publishers
-          )}
-        </div>
-        <div>{EditionCardUtils.getWorldCatElem(instance)}</div>
-        <Link to="/license">{EditionCardUtils.getLicense(previewItem)}</Link>
-      </CardContent>
-      <CardActions display="flex" flexDir="column" whiteSpace="nowrap" gap={4}>
-        {EditionCardUtils.getCtas(
-          previewItem,
-          instance.title,
-          !!cookies[NYPL_SESSION_ID]
-        )}
-      </CardActions>
-    </Card>
+      <Flex gap="xs" flexDirection={{ base: "column", md: "row" }}>
+        {isLoginRequired && <CardRequiredBadge />}
+        {isFeaturedEdition && <FeaturedEditionBadge />}
+      </Flex>
+      <Card
+        imageProps={{
+          src: EditionCardUtils.getCover(edition.links),
+          size: "xsmall",
+          aspectRatio: "original",
+          alt: ``,
+        }}
+        layout="row"
+        isAlignedRightActions
+        id={`card-${instance.instance_id}`}
+        paddingTop="m"
+      >
+        <CardHeading
+          level="h3"
+          size="heading6"
+          sx={{ span: { fontSize: "18px" } }}
+        >
+          <Flex alignItems="center">
+            <span>
+              {edition.publication_date
+                ? edition.publication_date
+                : "Edition Year Unknown"}
+            </span>
+            {isPhysicalEdition && <PhysicalEditionBadge />}
+          </Flex>
+        </CardHeading>
+        <CardContent>
+          <div>
+            <PublisherAndLocation
+              pubPlace={instance.publication_place}
+              publishers={instance.publishers}
+            />
+          </div>
+          <WorldCat instance={instance} />
+          <Link to="/license">{EditionCardUtils.getLicense(previewItem)}</Link>
+          {isPhysicalEdition && <ScanAndDeliverBlurb />}
+          {isUniversityPress && <UpBlurb publishers={edition.publishers} />}
+        </CardContent>
+        <CardActions
+          display="flex"
+          flexDir="column"
+          whiteSpace="nowrap"
+          gap="xs"
+        >
+          <Ctas item={previewItem} title={instance.title} />
+        </CardActions>
+      </Card>
+    </Box>
   );
 };